Now throw'ing instead of return'ing in the begin section. This actually stops the script and prevents it from going into the "process" section. Otherwise, it would continue to "process" even when it couldn't find ffmpeg.
Added estimated time remaining in pre-encode and encode step. This is part of the activity message, rather than using Write-Progress -SecondsRemaining as -SecondsRemaining literally only uses seconds, so for long encodes a user would have to convert this to hours/minutes manually; how silly. Instead we do the conversion to hours/minutes and seconds for them.